+++ Libplanet.Tests/Action/Sys/MintTest.cs
@@ -12,7 +12,7 @@ namespace Libplanet.Tests.Action.Sys
     public class MintTest
     {
         // ReSharper disable once InconsistentNaming
-        private static readonly Currency FOO = Currency.Uncapped("FOO", 2, null);
+        private static readonly Currency FOO = Currency.Capped("FOO", 2, (1000000, 0), null);
 
         // ReSharper disable once InconsistentNaming
         private static readonly Currency BAR = Currency.Uncapped("BAR", 0, null);
@@ -68,6 +68,7 @@ namespace Libplanet.Tests.Action.Sys
             var prevStates = new AccountStateDeltaImpl(
                 accountStateGetter: addr => new IValue[addr.Count],
                 accountBalanceGetter: (addr, c) => c * 0,
+                totalSupplyGetter: c => null,
                 signer: signer
             );
             BlockHash genesisHash = random.NextBlockHash();
@@ -91,6 +92,9 @@ namespace Libplanet.Tests.Action.Sys
             Assert.Equal(BAR * 0, nextStates.GetBalance(signer, BAR));
             Assert.Equal(FOO * 123456, nextStates.GetBalance(recipient, FOO));
             Assert.Equal(BAR * 0, nextStates.GetBalance(recipient, BAR));
+
+            var mintSupplyOverflowFoo = new Mint(recipient, FOO.MaximumSupply.Value * 2);
+            Assert.Throws<SupplyOverflowException>(() => mintSupplyOverflowFoo.Execute(context));
 
             var mintBar = new Mint(recipient, BAR * 10);
             NonNativeTokenException exc = Assert.Throws<NonNativeTokenException>(